Safari-611.1.10/LayoutTests/accessibility/mac/input-title-expected.txt


This tests that the input title is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS green.description is 'AXDescription: green'
PASS red.description is 'AXDescription: red'
PASS successfullyParsed is true

TEST COMPLETE
